Move all the build steps into its own service
